# uploads/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SpinalCore Graph API",
    "version": "1.0.0",
    "description": "Welcome to the reference documentation for the Spinalcore Digital Twin REST API! </br></br>To learn about the common use cases and concept of Spinalcore REST APIs, check out our resource center  (https://resourcecenter.fr.spinalcom.com/ressources-d%C3%A9veloppeur).</br></br>In addition to Spinalcore API Reference, we also provide complete documentation for integrator that need to install and implement Spinalcore Digital Twin Operating System on their built environment (https://resourcecenter.fr.spinalcom.com/ressources-int%C3%A9grateur)",
    "termsOfService": "",
    "contact": {
      "email": "developers@spinalcom.com"
    },
    "x-logo": {
      "url": "/logo.png"
    },
    "x-preferred": true,
    "x-providerName": "spinalcom.com",
    "x-unofficialSpec": true
  },
  "components": {
    "securitySchemes": {
      "OauthSecurity": {
        "type": "oauth2",
        "description": "This API uses OAuth 2 with the implicit grant flow. [More info](https://api.example.com/docs/auth)",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "http://localhost:8080/oauth2/token",
            "scopes": {
              "read": "read request",
              "admin": "admin request"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/node/{id}/read": {
      "get": {
        "security": [
          {
            "OauthSecurity": ["read"]
          }
        ],
        "description": "Return node object with parent and children relation",
        "summary": "Gets Node",
        "tags": ["Nodes"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "use the dynamic ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Node"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
...
    "/api/v1/node/{id}/download_file": {
      "post": {
        "security": [
          {
            "OauthSecurity": ["read"]
          }
        ],
        "description": "Download a Doc",
        "summary": "Download a Doc",
        "tags": ["Nodes"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "use the dynamic ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Download Successfully"
          },
          "400": {
            "description": "Download not Successfully"
          }
        }
      }
    },
...
    "/api/v1/room/{id}/control_endpoint_list": {
      "get": {
        "security": [
          {
            "OauthSecurity": ["read"]
          }
        ],
        "description": "Return list of control endpoint",
        "summary": "Gets a list of control endpoint",
        "tags": ["Geographic Context"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "use the dynamic ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/EndPointRoom"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
...
    "/api/v1/workflow/{id}/update": {
      "put": {
        "security": [
          {
            "OauthSecurity": ["read"]
          }
        ],
        "description": "update the workflow",
        "summary": "update the workflow",
        "tags": ["Workflow & ticket"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "use the dynamic ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["newNameWorkflow"],
                "properties": {
                  "newNameWorkflow": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
...
    "/api/v1/roomsGroup/{contextId}/category/{categoryId}/group_list": {
      "get": {
        "security": [
          {
            "OauthSecurity": ["read"]
          }
        ],
        "description": "Return list of group roomsGroup",
        "summary": "Gets a list of group roomsGroup",
        "tags": ["Rooms Group"],
        "parameters": [
          {
            "in": "path",
            "name": "contextId",
            "description": "use the dynamic ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "path",
            "name": "categoryId",
            "description": "use the dynamic ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/CategoryEvent"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Node": {
      "type": "object",
      "properties": {
        "dynamicId": {
          "type": "integer"
        },
        "staticId": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "children_relation_list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Relation"
          }
        }
      }
    },
    "Relation": {
      "type": "object",
      "properties": {
        "dynamicId": {
          "type": "integer"
        },
        "staticId": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "children_number": {
          "type": "integer"
        }
      }
    },
...
    "EndPointRoom": {
      "type": "object",
      "properties": {
        "dynamicId": {
          "type": "integer"
        },
        "staticId": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "currentValue": {
          "type": "integer"
        }
      }
    },
...
    "CategoryEvent": {
      "type": "object",
      "properties": {
        "dynamicId": {
          "type": "integer"
        },
        "staticId": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "type": {
          "type": "string"
        }
      }
    },
...
  },
...
}